requirements.txt
rplidar-roboticia
matplotlib
numpy
pandas
plotly